Move bootstrap peer list and connectToBootstrapPeers to p2p/bootstrap.go
